Adds MYSSHKEY variable to make experiment portable

+++ experiments/regression/run.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
-set -e -x
+set -ex
+
+if [ -z "$MYSSHKEY" ]; then
+  echo "Expecting MYSSHKEY variable"
+  exit 1
+fi
 
 outdir="results/benchoutput"
+factsdir="results/facts"
 
 # delete previous results
 rm -fr $outdir/*
+rm -fr $factsdir/*
 
 # get repetitions
 reps=`cat vars.yml | grep 'repetitions:' | awk '{print $2}'`
@@ -13,21 +20,24 @@ if [ -z "$reps" ]; then
   exit 1
 fi
 
+docker pull ivotron/ansible:2.2.0.0
+
+echo '' > ansible/ansible.log
+
 for i in `seq 1 $reps` ; do
   mkdir -p $outdir/repetition/$i
-  docker run --rm -ti \
-    -v `pwd`/ansible:/experiment \
-    -v `pwd`/vars.yml:/experiment/vars.yml \
-    -v `pwd`/$outdir/repetition/$i:/results \
-    -v $SSH_AUTH_SOCK:/ssh-agent \
-    -e SSH_AUTH_SOCK=/ssh-agent \
-    --workdir=/experiment \
+  docker run --rm \
+    -v `pwd`:/experiment \
+    -v $MYSSHKEY:/root/.ssh/id_rsa \
+    --workdir=/experiment/ansible \
     --net=host \
     --entrypoint=/bin/bash \
     ivotron/ansible:2.2.0.0 -c \
       "ansible-playbook \
-        -e @vars.yml \
-        -e local_results_path=/results \
+        -e @/experiment/vars.yml \
+        -e local_results_path=/experiment/$outdir/repetition/$i \
         playbook.yml"
-    mv $outdir/repetition/$reps/facts results/
 done
+
+mv $outdir/repetition/$reps/facts/* $factsdir
+rm -fr $outdir/repetition/*/facts
